backend/WorkTrack/requirements.txt
api==0.0.7
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.9.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
idna==3.10
nose==1.3.7
openpyxl==3.1.5
psycopg2-binary==2.9.11
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
